Add files via upload
 Shelter Animal Adoption Predictor

One of the biggest challenges of animal shelters is the allocation of ~6.5 million rescued animals. Unfortunately, ~1.5 million rescued animals are euthanized every year. Our goal is to develop a model to predict which animals have a higher probability of being adopted at shelters. In determining this, efforts may be more properly allocated to increase the number of adoptions and decrease the number of euthanizations at shelters. 

We aim to gain insight on which pet attributes contribute to the likelihood of adoption and their significance. In turn, the model outcome may help shelters identify which animals have a smaller chance of being adopted and dedicate greater effort to find these animals a home. Additionally, this predictor model may aid in providing an estimation on which and how many animals will be adopted, which can help with more effective allocation of resources (ie. space, equipment, etc) and give revenue projections.